requirements.txt
asyncio
Playwright
python-dotenv